Cleaning up branch for pull request
Renamed some files, added comments, made a couple things less dumb
